Refactored MediaProviderMusicTest splitting it up into separate classes
